feat: Add spinning donut demo with optimized rendering and color handling
