+++ kernel/src/graphics/native_pixel.hpp
@@ -3,6 +3,7 @@
 
 #include <defines.hpp>
 #include <types.hpp>
+#include "graphics/color.hpp"
 
 namespace Graphics
 {
@@ -21,6 +22,22 @@ struct PACK NativePixel {
 
     bool operator==(const NativePixel &other) const = default;
     explicit constexpr operator u32() const { return value; }
+
+    /**
+     * @brief Converts a logical RGBA color to a hardware-specific NativePixel.
+     */
+    NODISCARD static constexpr NativePixel FromColor(const Color &c, const PixelFormat &pf)
+    {
+        const u32 r_mask = (1 << pf.red_mask_size) - 1;
+        const u32 g_mask = (1 << pf.green_mask_size) - 1;
+        const u32 b_mask = (1 << pf.blue_mask_size) - 1;
+
+        u32 raw = (static_cast<u32>(c.r & r_mask) << pf.red_pos) |
+                  (static_cast<u32>(c.g & g_mask) << pf.green_pos) |
+                  (static_cast<u32>(c.b & b_mask) << pf.blue_pos);
+
+        return NativePixel(raw);
+    }
 };
 
 static_assert(sizeof(NativePixel) == sizeof(u32), "NativePixel must be 32-bit");
